requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
alabaster==1.0.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
certifi==2024.8.30
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==4.2.16
django-filter==24.3
django-libsql==0.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
drf-yasg==1.21.8
frozenlist==1.4.1
gunicorn==23.0.0
idna==3.10
imagesize==1.4.1
inflection==0.5.1
Jinja2==3.1.4
libsql-client==0.3.1
libsql-experimental==0.0.40
MarkupSafe==2.1.5
multidict==6.1.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
snowballstemmer==2.2.0
Sphinx==8.0.2
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.7.0
yarl==1.12.1
